% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{\%if_na\%}
\alias{\%if_na\%}
\title{infix function for replacing NA or NULL}
\usage{
a \%if_na\% b
}
\description{
This function replaces possible NA or NULL in a single line of code.
}
\examples{
NA \%if_na\% iris
}
\keyword{NA}
\keyword{replace}
