% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{\%if_null\%}
\alias{\%if_null\%}
\title{infix function for replacing NULL}
\usage{
a \%if_null\% b
}
\description{
This function replaces possible NULL in a single line of code.
}
\examples{
NULL \%if_null\% iris
}
\keyword{null}
\keyword{replace}
