% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{named_group_split}
\alias{named_group_split}
\title{dplyr extension for group_split}
\usage{
named_group_split(...)
}
\description{
This function returns the list from executing dplyr::group_split with its names as the group.
}
\examples{
if(require(dplyr)){ identical(iris \%>\% named_group_split(c('Species','Sepal.Length')),iris \%>\% named_group_split(Species,Sepal.Length)) }
}
\keyword{NA}
\keyword{replace}
